pyproject.toml
[project]
name = "messager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi>=0.133.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "redis>=7.2.1",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.41.0",
    "websockets>=16.0",
]

[dependency-groups]
dev = [
    "types-pyjwt>=1.7.1",
    "types-redis>=4.6.0.20241004",
]
